[refactor] fixed folder path
