+++ application/__init__.py
@@ -1,4 +1,1 @@
-from application import application
-from .image import Image
-from .parser import Parser
-from application import mymongo
+from .app import create_app
