+++ application/api/resources/__init__.py
@@ -1,0 +1,2 @@
+from .images import Images
+from .images_list import ImagesList
