+++ application/api/resources/images_list.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource
 # TODO: db connection should be one for all blueprints
-from application.mymongo import get_all_images
+from application.image_bot.mymongo import get_all_images
 
 
 class ImagesList(Resource):
