+++ application/image_bot/tests/mocks/mock_mymongo.py
