+++ application/image_bot/tests/test_mymongo.py
@@ -1,7 +1,7 @@
 import unittest
 
 from PIL import Image
-from application.tests.mocks.mock_mymongo import get_image_by_hash, get_all_images_count, save
+from application.image_bot.tests.mocks.mock_mymongo import get_image_by_hash, get_all_images_count
 import imagehash
