+++ application/image_scrapy/image_scrapy/items.py
