+++ application/image_scrapy/image_scrapy/pipelines.py
