+++ application/image_scrapy/image_scrapy/settings.py
