+++ application/image_scrapy/image_scrapy/spiders/__init__.py
