+++ application/image_scrapy/image_scrapy/spiders/image_spider.py
