+++ application/image_scrapy/scrapy.cfg
