+++ application/logger/__init__.py
@@ -1,0 +1,1 @@
+from .logger import Logger
