+++ application/settings.py
@@ -4,3 +4,7 @@ DB = 'test'
 COLLECTION = 'image'
 HOST = 'localhost'
 PORT = 27017
+
+# STATIC
+
+STATIC_PATH = '/static'
