+++ main.py
@@ -1,54 +1,10 @@
-from flask import Flask, Response, render_template, request, redirect
-from flask_socketio import SocketIO, emit
+from application import create_app
 
-from application import application
-from api import api
-from logger import Logger
-
-# app init
-app = Flask(__name__, static_url_path='/static')
-socketio = SocketIO(app)
-logger = Logger(socketio)
-   
-# default route
-@app.route('/')
-def index():
-    return render_root_template()
-
-
-@app.route('/parse', methods=['GET', 'POST'])
-def post():
-    if request.method == 'POST':
-        # TODO : check is True url / don't believe user
-        print(request.form)
-        application.parse(request.form['url'], request.form['quantity'])
-        return redirect('/parse')
-    return render_template("parse.html")
-
-
-@app.route('/about', methods=['GET', 'POST'])
-def about():
-    if request.method == 'GET':
-         return render_template("about.html")
-
-
-@app.route("/download/image")
-def download_image():
-    logger.send_message('download completed')
-    return Response(application.get_images_as_xml(), mimetype="text/xml",
-                    headers={"Content-disposition": "attachment;"})
-
-
-app.register_blueprint(api, url_prefix='/api')
-
-# helpers
-def render_root_template():
-    return render_template('index.html', images_count=application.get_images_count())
-
+app = create_app()
 
 if __name__ == '__main__':
     from gevent import pywsgi
     from geventwebsocket.handler import WebSocketHandler
-    app.config['DEBUG'] = True
-    server = pywsgi.WSGIServer(('', 5000), app, handler_class=WebSocketHandler)
+    print('Open in browser - 0.0.0.0:8080')
+    server = pywsgi.WSGIServer(('0.0.0.0', 8080), app, handler_class=WebSocketHandler)
     server.serve_forever()
